requirements.txt
pytest
coverage
pytest-cov
ruff
fastapi
uvicorn
jinja2
httpx
aiohttp
python-binance
pedantic
datetime
typing
matplotlib
mplfinance
beautifulsoup4
